backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pandas==2.1.4
openpyxl==3.1.2
aiosqlite==0.21.0
httpx==0.28.1
python-dotenv==1.1.0
